requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
alpaca-py==0.43.2
annotated-types==0.7.0
attrs==26.1.0
beautifulsoup4==4.14.3
bitarray==3.8.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
ckzg==2.1.7
cryptography==46.0.6
curl_cffi==0.13.0
cytoolz==1.1.0
eth-account==0.13.7
eth-hash==0.8.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==6.0.0
eth-utils==5.3.1
eth_abi==6.0.0b1
frozendict==2.4.7
frozenlist==1.8.0
hexbytes==1.3.1
hyperliquid-python-sdk==0.22.0
idna==3.11
msgpack==1.1.2
multidict==6.7.1
multitasking==0.0.12
numpy==2.4.4
pandas==3.0.2
parsimonious==0.10.0
peewee==4.0.3
platformdirs==4.9.4
propcache==0.4.1
protobuf==7.34.1
pycares==5.0.1
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
regex==2026.3.32
requests==2.33.1
rlp==4.1.0
setuptools==82.0.1
six==1.17.0
soupsieve==2.8.3
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websocket-client==1.9.0
websockets==16.0
yarl==1.23.0
yfinance==1.2.0
